requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
psycopg2-binary==2.9.9
django-redis==5.4.0
django-ratelimit==4.1.0
whitenoise==6.6.0
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1